Implemented Levels in Review Mode
Added a lot of code which pretty much implements a level system in review mode. No level select for review which can be added if wanted but as of now, it begins at level 1.
Added new files under a new hidden folder for words to review for each level. Can be implemented differently but I can't think of a more efficient way right now